+++ sandbox/chat-notification-pubnub-example/facade/src/permission-key.enum.ts
@@ -18,5 +18,5 @@ export enum PermissionKey {
   ViewChannel = 'ViewChannel',
   CreateChannel = 'CreateChannel',
   UpdateChannel = 'UpdateChannel',
-  DeleteChannel = 'DeleteChannel'
+  DeleteChannel = 'DeleteChannel',
 }