+++ sandbox/chat-notification-pubnub-example/services/notifications-service/src/datasources/notifdb.datasource.ts
@@ -11,7 +11,7 @@ const config = {
   user: process.env.NOTIF_DB_USER,
   password: process.env.NOTIF_DB_PASSWORD,
   database: process.env.NOTIF_DB_DATABASE,
-  schema: 'main'
+  schema: 'main',
 };
 
 // Observe application's life cycle to disconnect the datasource when
@@ -19,8 +19,10 @@ const config = {
 // gracefully. The `stop()` method is inherited from `juggler.DataSource`.
 // Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
 @lifeCycleObserver('datasource')
-export class NotifdbDataSource extends juggler.DataSource
-  implements LifeCycleObserver {
+export class NotifdbDataSource
+  extends juggler.DataSource
+  implements LifeCycleObserver
+{
   static dataSourceName = NotifDbSourceName;
   static readonly defaultConfig = config;