+++ services/authentication-service/src/__tests__/fixtures/datasources/auth-mock-data.datasource.ts
@@ -1,6 +1,12 @@
 import {inject} from '@loopback/core';
 import {juggler} from '@loopback/repository';
-import * as config from './auth-mock-data.datasource.json';
+
+const config = {
+  name: 'authMockData',
+  connector: 'memory',
+  localStorage: '',
+  file: '',
+};
 
 export class AuthMockDataDataSource extends juggler.DataSource {
   static dataSourceName = 'AuthDB';