+++ services/feature-toggle-service/src/component.ts
@@ -43,6 +43,11 @@ import {
 } from './providers';
 import {FeatureToggleSequence} from './sequence';
 import {IToggleServiceConfig} from './types';
+import {
+  FeaturesController,
+  ProjectsController,
+  StrategiesController,
+} from './controllers';
 
 export class FeatureToggleServiceComponent implements Component {
   constructor(
@@ -52,7 +57,6 @@ export class FeatureToggleServiceComponent implements Component {
     private readonly config?: IToggleServiceConfig,
   ) {
     this.bindings = [];
-    this.providers = {};
 
     // Mount core component
     this.application.component(CoreComponent);
@@ -82,14 +86,21 @@ export class FeatureToggleServiceComponent implements Component {
     ];
 
     this.models = [Features, Parameters, Projects, Strategies, StrategiesType];
+    this.providers = {
+      [StrategyBindings.FEATURE_FLAG_ACTION.key]: FeatureFlagActionProvider,
+      [StrategyBindings.METADATA.key]: FeatureFlagMetadataProvider,
+      [StrategyBindings.TENANT_FEATURE.key]: TenantFeatureProvider,
+      [StrategyBindings.USER_FEATURE.key]: UserFeatureProvider,
+      [StrategyBindings.SYSTEM_FEATURE.key]: SystemFeatureProvider,
+    };
+
+    this.controllers = [
+      FeaturesController,
+      ProjectsController,
+      StrategiesController,
+    ];
   }
-  providers?: ProviderMap = {
-    [StrategyBindings.FEATURE_FLAG_ACTION.key]: FeatureFlagActionProvider,
-    [StrategyBindings.METADATA.key]: FeatureFlagMetadataProvider,
-    [StrategyBindings.TENANT_FEATURE.key]: TenantFeatureProvider,
-    [StrategyBindings.USER_FEATURE.key]: UserFeatureProvider,
-    [StrategyBindings.SYSTEM_FEATURE.key]: SystemFeatureProvider,
-  };
+  providers?: ProviderMap = {};
 
   bindings?: Binding[] = [];
   /**