+++ sandbox/pubnub-example/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.2.4 (2021-12-17)</small>
+
+* fix(core): upgrade all deps (#438) ([c52b7f1](https://github-personal/sourcefuse/loopback4-microservice-catalog/commit/c52b7f1)), closes [#438](https://github-personal/sourcefuse/loopback4-microservice-catalog/issues/438)
+
+
+
+
+
+## <small>0.2.3 (2021-11-11)</small>
+
+**Note:** Version bump only for package @sourceloop/pubnub-example
+
+
+
+
+
+## <small>0.2.2 (2021-11-09)</small>
+
+**Note:** Version bump only for package @sourceloop/pubnub-example
+
+
+
+
+
+## <small>0.2.1 (2021-11-09)</small>
+
+**Note:** Version bump only for package @sourceloop/pubnub-example
+
+
+
+
+
 ## 0.2.0 (2021-10-24)
 
 * feat(search-service): add search service, with a sandbox example (#341) ([501513c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/501513c)), closes [#341](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/341)