+++ services/search-service/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.5.0 (2022-01-10)
+
+* feat(search-service): add option to change source name (#450) ([b917898](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/b917898)), closes [#450](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/450) [#449](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/449)
+
+
+
+
+
+## <small>0.4.1 (2021-12-26)</small>
+
+* fix(core): fix versions in package locks and update references in tsconfig ([36b563e](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/36b563e)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+
+
+
+
+
+## 0.4.0 (2021-12-26)
+
+* feat(core): add support for node.js v17 and upgrade to @loopback/cli v3 ([4115b6c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/4115b6c)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+* fix(docs): add badge for node and npm version ([caaa1bd](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/caaa1bd))
+
+
+### BREAKING CHANGE
+
+* drop support for node.js v10
+
+
+
+
+## 0.3.0 (2021-12-17)
+
+* fix(core): upgrade all deps (#438) ([c52b7f1](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/c52b7f1)), closes [#438](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/438)
+* fix(docs): added badges ([a66ed14](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/a66ed14))
+* feat(sandbox): add search-client example (#427) ([4735004](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/4735004)), closes [#427](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/427) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+
+
+### BREAKING CHANGE
+
+* bump major versions
+
+
+
+
+## <small>0.2.7 (2021-11-11)</small>
+
+**Note:** Version bump only for package @sourceloop/search-service
+
+
+
+
+
+## <small>0.2.6 (2021-11-09)</small>
+
+* refactor(authentication-service): version bump in all services for release (#418) ([0808265](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/0808265)), closes [#418](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/418) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+
+
+
+
+
+## <small>0.2.4 (2021-11-09)</small>
+
+* fix(search-client): reduce gap between input box and recent searches  (#400) ([2e5849d](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/2e5849d)), closes [#400](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/400)
+
+
+
+
+
 ## <small>0.2.3 (2021-10-29)</small>
 
 * fix(search-service): handle case in which match has whitespaces (#395) ([50e467e](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/50e467e)), closes [#395](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/395)